+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/Berries.java
@@ -5,7 +5,6 @@ import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.iguanatweaksreborn.module.misc.DataPacks;
 import insane96mcp.iguanatweaksreborn.setup.ISORegistries;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
@@ -24,30 +23,21 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.registries.RegistryObject;
 
-@Label(name = "Berries")
 @LoadFeature(module = Modules.Ids.WORLD)
 public class Berries extends Feature {
 
     public static final RegistryObject<BlockItem> SWEET_BERRY_SEEDS = ISORegistries.ITEMS.register("sweet_berry_seeds", () -> new ItemNameBlockItem(Blocks.SWEET_BERRY_BUSH, new Item.Properties()));
 
-    @Config
-    @Label(name = "No damage if dressed", description = "Berry bushes no longer deal damage when walking in them with leggings and boots")
+    @Config(description = "Berry bushes no longer deal damage when walking in them with leggings and boots")
     public static Boolean noDamageIfDressed = true;
 
-    @Config
-    @Label(name = "Data pack", description = """
+    @Config(description = """
 		Makes sweet berries not plantable, requiring seeds, and also enables a data pack that makes the following changes:
 		* Makes sweet berry bushes drop seeds
 		* Makes sweet berry patches has less plants and plants will not be all grown up
 		If berry good is present, a different datapack is enabled that integrates with the mod
 	""")
     public static Boolean dataPack = true;
-
-    /*@Config
-    @Label(name = "Data pack Berry Good", description = """
-		Same as 'Data pack' but uses Berry Good seeds
-	""")
-    public static Boolean dataPackBerryGood = false;*/
 
     public Berries(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);