+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/Nether.java
@@ -5,7 +5,6 @@ import insane96mcp.iguanatweaksreborn.data.generator.ISOBlockTagsProvider;
 import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.iguanatweaksreborn.module.misc.DataPacks;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
@@ -24,26 +23,20 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModList;
 import org.apache.commons.lang3.mutable.MutableInt;
 
-@Label(name = "Nether")
 @LoadFeature(module = Modules.Ids.WORLD)
 public class Nether extends Feature {
-    public static final String REQUIRES_CORNERS_LANG = InsaneSO.MOD_ID + ".requires_corners";
+    public static final String REQUIRES_CORNERS_LANG = InsaneSO.lang("requires_corners");
     public static final TagKey<Block> PORTAL_CORNERS = ISOBlockTagsProvider.create("portal_corners");
 
-    @Config
-    @Label(name = "Disable Nether Roof and reduce block ratio", description = "Enables a data pack that makes the nether 128 blocks high instead of 256, effectively disabling the \"Nether Roof\" and reduces the 8 block ratio between nether and other dimensions to 4.")
+    @Config(name = "Disable Nether Roof and reduce block ratio", description = "Enables a data pack that makes the nether 128 blocks high instead of 256, effectively disabling the \"Nether Roof\" and reduces the 8 block ratio between nether and other dimensions to 2.")
     public static Boolean netherTweaks = true;
-    @Config
-    @Label(name = "Remove Lava Pockets", description = "If true, lava pockets in the nether are removed. If quark is installed, this is disabled")
+    @Config(description = "If true, lava pockets in the nether are removed. If quark is installed, this is disabled")
     public static Boolean removeLavaPockets = true;
-    @Config
-    @Label(name = "Disable Nether", description = "If true you cannot lit nether portals")
+    @Config(description = "If true you cannot lit nether portals")
     public static Boolean disableNether = false;
-    @Config
-    @Label(name = "Portal wait time", description = "How much time inside the portal to pass it. Vanilla is 80 ticks")
+    @Config(description = "How much time inside the portal to pass it. Vanilla is 80 ticks")
     public static Integer portalWaitTime = 50;
-    @Config
-    @Label(name = "Portal requires Crying Obsidian", description = "The portal requires Crying Obsidian in the corners to turn it on (in the overworld). The block tag 'iguanatweaksreborn:portal_corners' can be used to change the required blocks for the corners")
+    @Config(description = "The portal requires Crying Obsidian in the corners to turn it on (in the overworld). The block tag 'iguanatweaksreborn:portal_corners' can be used to change the required blocks for the corners")
     public static Boolean portalRequiresCryingObsidian = true;
 
     public Nether(Module module, boolean enabledByDefault, boolean canBeDisabled) {